[#66] Refactor : 가독성 향상 위해 파일명 변경
